add comments and scenario

+++ observable.h
@@ -2,12 +2,14 @@
 #define _OBSERVABLE_
 #include "observer.h"
 #include <vector>
+// We use the vector class of STL because it's more intuitive to
+// manipulate than the other alternatives like the list templates
 
 class Observable
 {
 public:
-	double Tconsigne;
-	double Tmeasured;
+	int Tconsigne;
+	int Tmeasured;
 	std::vector<Observer *> observer_list;
 	virtual ~Observable(){};
 	virtual void measure(int T) = 0;
